backend/requirements.txt
ibm-watsonx-ai>=1.4.0
astrapy>=2.0.0
docling>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0